Feat: #46 우주선 애니메이션 적용
